custom-elements/registering-custom-elements/unresolved-element-pseudoclass/unresolved-element-pseudoclass-matching-query-selector.html

<!DOCTYPE html>
<html>
<head>
<title>The :unresolved pseudoclass matching with Document.querySelector()</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<meta name="assert" content="The :unresolved pseudoclass must match all custom elements whose created callback has not yet been invoked">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#unresolved-element-pseudoclass">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    doc.body.innerHTML = '<x-a></x-a>';
    var customElement = doc.querySelector(':unresolved');

    assert_not_equals(customElement, null,
        'Unregistered custom element should be accessible by :unresolved');
    assert_equals(customElement.localName, 'x-a', 'Wrong custom element');
}, 'Document.querySelector(): Unresolved custom element should be accessible by :unresolved');


test(function() {
    var doc = newHTMLDocument();
    doc.body.innerHTML = '<x-b></x-b>';
    var customElement = doc.querySelector(':unresolved');
    assert_not_equals(customElement, null,
        'Unresolved custom element should be accessible by :unresolved');
    assert_equals(customElement.localName, 'x-b', 'Wrong custom element');

    doc.registerElement('x-b');
    customElement = doc.querySelector(':unresolved');
    assert_equals(customElement, null,
        'Resolved custom element should not be accessible by :unresolved');
}, 'Document.querySelector(): Resolved custom element should not be accessible by :unresolved');


test(function() {
    var doc = newHTMLDocument();
    doc.body.innerHTML = '<x-c></x-c><x-d></x-d>';
    var customElement = doc.querySelector(':unresolved');
    assert_not_equals(customElement, null,
        'Unregistered custom element should be accessible by :unresolved');
    assert_equals(customElement.localName, 'x-c', 'Wrong custom element');

    doc.registerElement('x-c');
    customElement = doc.querySelector(':unresolved');
    assert_not_equals(customElement, null,
        'Unresolved custom element should be accessible by :unresolved');
    assert_equals(customElement.localName, 'x-d', 'Wrong custom element');
}, 'Document.querySelector(): Unresolved custom element should be accessible by :unresolved. ' +
    'There are several unresolved elements');


test(function() {
    var doc = newHTMLDocument();
    var customElement = doc.createElement('x-e');
    doc.body.appendChild(customElement);
    var queryResult = doc.querySelector(':unresolved');

    assert_not_equals(queryResult, null,
        'Unregistered custom element should be accessible by :unresolved');
    assert_equals(queryResult.localName, 'x-e', 'Wrong custom element');
}, 'Document.querySelector(): Unresolved custom element, created via Document.createElement(), '+
    'should be accessible by :unresolved');


test(function() {
    var doc = newHTMLDocument();
    doc.body.innerHTML = '<div><x-f></x-f><div>';
    var customElement = doc.querySelector(':unresolved');

    assert_not_equals(customElement, null,
        'Unregistered custom element should be accessible by :unresolved');
    assert_equals(customElement.localName, 'x-f', 'Wrong custom element');
}, 'Document.querySelector(): Unresolved custom element inside div element '+
    'should be accessible by :unresolved');


test(function() {
    var doc = newHTMLDocument();
    var divElement = doc.createElement('div');
    doc.body.appendChild(divElement);
    divElement.innerHTML = '<a is="x-g"></a>';
    var queryResult = doc.querySelector(':unresolved');

    assert_not_equals(queryResult, null,
        'Unregistered custom element should be accessible by :unresolved');
    assert_equals(queryResult.localName, 'a', 'Wrong custom element');
}, 'Document.querySelector(): Unresolved custom element, inside div element, ' +
    'both created via Document.createElement(), should be accessible by :unresolved');


testInIFrame('../../resources/x-element.html', function(doc) {
    var customElement = doc.querySelector(':unresolved');

    assert_not_equals(customElement, null,
        'Unregistered custom element should be accessible by :unresolved');
    assert_equals(customElement.localName, 'x-element', 'Wrong custom element');
}, 'Document.querySelector(): Unresolved custom element should be accessible by :unresolved '+
    'in loaded document');
</script>
</body>
</html>
